lambda_functions/requirements.txt
python-decouple==3.8
requests==2.32.3